+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - **ENH: Force `run_in_conda_env` function to launch the input executable in the same environment with `conda_env_name="self"`**
 - FIX: Make sure arrays have a CRS when reprojected with RPCs. If not, set `epsg:4326` as default, by convention.
+- FIX: Add an error if the window doesn't cover the dataset in `get_window`
 
 ## 1.48.0 (2025-11-07)